Isolate pre-yarn process helpers for check-package-manager
