+++ build-system/common/exec.js
@@ -21,20 +21,10 @@
 
 const childProcess = require('child_process');
 const log = require('fancy-log');
+const {spawnProcess, getOutput, getStdout, getStderr} = require('./process');
 const {yellow} = require('ansi-colors');
 
 const shellCmd = process.platform == 'win32' ? 'cmd' : '/bin/bash';
-
-/**
- * Spawns the given command in a child process with the given options.
- *
- * @param {string} cmd
- * @param {?Object} options
- * @return {!Object}
- */
-function spawnProcess(cmd, options) {
-  return childProcess.spawnSync(cmd, {shell: shellCmd, ...options});
-}
 
 /**
  * Executes the provided command with the given options, returning the process
@@ -104,42 +94,6 @@ function execOrThrow(cmd, msg) {
     throw error;
   }
   return p;
-}
-
-/**
- * Executes the provided command, returning the process object.
- * @param {string} cmd
- * @param {?Object} options
- * @return {!Object}
- */
-function getOutput(cmd, options = {}) {
-  const p = spawnProcess(cmd, {
-    'cwd': options.cwd || process.cwd(),
-    'env': options.env || process.env,
-    'stdio': options.stdio || 'pipe',
-    'encoding': options.encoding || 'utf-8',
-  });
-  return p;
-}
-
-/**
- * Executes the provided command, returning its stdout.
- * @param {string} cmd
- * @param {?Object} options
- * @return {string}
- */
-function getStdout(cmd, options) {
-  return getOutput(cmd, options).stdout;
-}
-
-/**
- * Executes the provided command, returning its stderr.
- * @param {string} cmd
- * @param {?Object} options
- * @return {string}
- */
-function getStderr(cmd, options) {
-  return getOutput(cmd, options).stderr;
 }
 
 module.exports = {
